requirements.txt
pyTelegramBotAPI==4.28.0
flake8==7.3.0
python-dotenv==1.1.1
requests==2.32.4
